api: Move events functions to `events` sub-directory
No functionality change, just for better organization of the code.
