+++ src/api/events/pollForEvents.js
@@ -1,6 +1,6 @@
 /* @flow strict-local */
-import type { Auth } from './apiTypes';
-import { apiGet } from './apiFetch';
+import type { Auth } from '../apiTypes';
+import { apiGet } from '../apiFetch';
 
 /** See https://zulipchat.com/api/get-events-from-queue */
 export default (auth: Auth, queueId: number, lastEventId: number) =>
