+++ src/api/events/registerForEvents.js
@@ -1,6 +1,6 @@
 /* @flow strict-local */
-import type { Auth, Narrow } from './apiTypes';
-import { apiPost, objectToParams } from './apiFetch';
+import type { Auth, Narrow } from '../apiTypes';
+import { apiPost, objectToParams } from '../apiFetch';
 
 type RegisterForEventsParams = {
   apply_markdown?: boolean,
