+++ src/api/index.js
@@ -9,8 +9,8 @@ import getTopics from './getTopics';
 import toggleMessageStarred from './messages/toggleMessageStarred';
 import typing from './typing';
 import getAlertWords from './alert_words/getAlertWords';
-import pollForEvents from './pollForEvents';
-import registerForEvents from './registerForEvents';
+import pollForEvents from './events/pollForEvents';
+import registerForEvents from './events/registerForEvents';
 import uploadFile from './uploadFile';
 import emojiReactionAdd from './emoji_reactions/emojiReactionAdd';
 import emojiReactionRemove from './emoji_reactions/emojiReactionRemove';
